move section UI to separate component
