+++ src/apps/throwdown/components/throwdown/section.jsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+function SectionTrigger( props ) {
+  const styles = {};
+  styles.fontWeight = props.playing ? 'bold' : 'normal';
+  styles.fontStyle = props.triggered ? 'italic' : 'normal';
+  const toggleTrigger = props.onSetTriggeredSection.bind(
+    null,
+    props.triggered ? null : props.slug
+  );
+  return (
+    <td>
+      <span
+        onClick={ toggleTrigger }
+        style={ styles }
+      >
+        { props.slug }
+      </span>
+    </td>
+  );
+}
+
+SectionTrigger.propTypes = {
+  playing: PropTypes.bool,
+  triggered: PropTypes.bool,
+  onSetTriggeredSection: PropTypes.func,
+  slug: PropTypes.string,
+  hue: PropTypes.number,
+};
+
+export default SectionTrigger;
